requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-multipart==0.0.22
reportlab==4.0.9
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
requests==2.31.0
pillow==10.3.0
